ob-core: Added option for attaching file from babel src block result
* ob-core.el (org-babel-format-result): Added `:results-param' option `attach' which when used with `file' will cause the resultant file path to be moved to the `org-attach-dir' for the node.
(org-babel-result-to-file): Added optional 3rd parameter `type' which will be used in place of "file" in link type, if present.
(org-babel-result-to-file-attachment): New function that moves the file at theh provided `result' path to the nodes attachment directory, creating it as necessary using `org-attach-dir'.  A new function was created to adhere to the design of the API and not complicated `org-babel-result-to-file' with the movement to the attachment directory.

* org-attach.el (org-attach-dir): Added autoload header to simplify dependencies necessary to support this feature (called in `org-babel-result-to-file-attachment').